# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost/Api-Livros'
    description: 'Servidor local'
paths:
...
  /usuario/foto:
    put:
      tags:
        - Usuário
      summary: 'Atualiza a foto de perfil do usuário autenticado'
      operationId: a76c0610241bf27126a5aaf1daaf901e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - url_foto
              properties:
                url_foto:
                  type: string
                  format: uri
                  example: 'https://meusite.com/imagens/foto.jpg'
              type: object
      responses:
        '200':
          description: 'Foto de perfil atualizada com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Foto de perfil atualizada com sucesso' }
                  foto_perfil: { type: string, format: uri, example: 'https://meusite.com/imagens/foto.jpg' }
                type: object
        '400':
          description: 'URL ausente ou inválida'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'A URL da foto é inválida' }
                type: object
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro ao atualizar a foto de perfil'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao atualizar a foto de perfil' }
                type: object
      security:
        -
          bearerAuth: []
  /usuario:
    get:
      tags:
        - Usuário
      summary: 'Lista os dados do usuário autenticado'
      operationId: 1ee74c0a21f0b1d380b76ab89163a811
      responses:
        '200':
          description: 'Usuário encontrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Usuário encontrado' }
                  usuario: { type: object, additionalProperties: true }
                type: object
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Usuário não encontrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Usuário não encontrado' }
                type: object
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Usuário
      summary: 'Edita os dados do usuário autenticado'
      operationId: 9096c93a3a91a723d09c1fb0c1f9db71
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                nome:
                  type: string
                  example: 'Luiz Felipe'
                email:
                  type: string
                  format: email
                  example: luiz@email.com
              type: object
      responses:
        '200':
          description: 'Usuário atualizado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Usuario atualizado com sucesso' }
                  usuario: { properties: { id_usuario: { type: integer, example: 1 } }, type: object }
                type: object
        '401':
          description: 'Não autenticado'
        '409':
          description: 'E-mail já está em uso por outro usuário'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'O email ja esta em uso por outro usuario' }
                type: object
        '500':
          description: 'Erro ao atualizar usuário'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao atualizar usuário' }
                type: object
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Usuário
      summary: 'Deleta o usuário autenticado'
      operationId: c22821ed743d1fe670a7ff3a999c1f5f
      responses:
        '200':
          description: 'Usuário deletado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Usuário deletado com sucesso' }
                type: object
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro ao deletar usuário'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao deletar usuário' }
                type: object
      security:
        -
          bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...